+++ baybe/recommenders/__init__.py
@@ -10,6 +10,7 @@ from baybe.recommenders.pure.bayesian.botorch import BotorchRecommender
 from baybe.recommenders.pure.bayesian.sequential_greedy import (
     SequentialGreedyRecommender,
 )
+from baybe.recommenders.pure.llm import LLMRecommender
 from baybe.recommenders.pure.nonpredictive.clustering import (
     GaussianMixtureClusteringRecommender,
     KMeansClusteringRecommender,
@@ -32,4 +33,5 @@ __all__ = [
     "SequentialGreedyRecommender",
     "SequentialMetaRecommender",
     "StreamingSequentialMetaRecommender",
+    "LLMRecommender",
 ]